+++ src/Toolkit/Identifiers/GuidTypeMap.php
@@ -12,25 +12,44 @@ declare(strict_types=1);
 
 namespace CloudCreativity\Modules\Toolkit\Identifiers;
 
-final readonly class GuidTypeMap
+use Ramsey\Uuid\UuidInterface;
+use UnitEnum;
+
+use function CloudCreativity\Modules\Toolkit\enum_string;
+
+final class GuidTypeMap
 {
     /**
      * GuidTypeMap constructor
      *
      * @param array<string, mixed> $map
      */
-    public function __construct(private array $map)
+    public function __construct(private array $map = [])
     {
+    }
+
+    /**
+     * Define an alias to type mapping.
+     *
+     * @param UnitEnum|string $alias
+     * @param UnitEnum|string $type
+     * @return void
+     */
+    public function define(UnitEnum|string $alias, UnitEnum|string $type): void
+    {
+        $alias = enum_string($alias);
+
+        $this->map[$alias] = $type;
     }
 
     /**
      * Get the GUID for the specified alias and id.
      *
-     * @param string $alias
-     * @param string|int $id
+     * @param UnitEnum|string $alias
+     * @param Uuid|UuidInterface|string|int $id
      * @return Guid
      */
-    public function guidFor(string $alias, string|int $id): Guid
+    public function guidFor(UnitEnum|string $alias, Uuid|UuidInterface|string|int $id): Guid
     {
         return Guid::make($this->typeFor($alias), $id);
     }
@@ -38,11 +57,13 @@ final readonly class GuidTypeMap
     /**
      * Get the GUID type for the specified alias.
      *
-     * @param string $alias
-     * @return string
+     * @param UnitEnum|string $alias
+     * @return UnitEnum|string
      */
-    public function typeFor(string $alias): string
+    public function typeFor(UnitEnum|string $alias): UnitEnum|string
     {
+        $alias = enum_string($alias);
+
         assert(
             isset($this->map[$alias]),
             sprintf('Alias "%s" is not defined in the type map.', $alias),
@@ -51,7 +72,7 @@ final readonly class GuidTypeMap
         $type = $this->map[$alias] ?? null;
 
         assert(
-            is_string($type) && !empty($type),
+            (is_string($type) || $type instanceof UnitEnum) && !empty($type),
             sprintf('Expecting type for alias "%s" to be a non-empty string.', $alias),
         );