+++ tests/Unit/Toolkit/Identifiers/GuidTypeMapTest.php
@@ -13,16 +13,17 @@ declare(strict_types=1);
 namespace CloudCreativity\Modules\Tests\Unit\Toolkit\Identifiers;
 
 use AssertionError;
+use CloudCreativity\Modules\Tests\TestBackedEnum;
+use CloudCreativity\Modules\Tests\TestBackedIntEnum;
+use CloudCreativity\Modules\Tests\TestUnitEnum;
 use CloudCreativity\Modules\Toolkit\Identifiers\Guid;
 use CloudCreativity\Modules\Toolkit\Identifiers\GuidTypeMap;
+use CloudCreativity\Modules\Toolkit\Identifiers\Uuid;
 use PHPUnit\Framework\Attributes\Depends;
 use PHPUnit\Framework\TestCase;
 
 class GuidTypeMapTest extends TestCase
 {
-    /**
-     * @return GuidTypeMap
-     */
     public function testItReturnsExpectedType(): GuidTypeMap
     {
         $map = new GuidTypeMap([
@@ -79,5 +80,25 @@ class GuidTypeMapTest extends TestCase
         $this->expectExceptionMessage('Alias "NotDefined" is not defined in the type map.');
 
         $map->typeFor('NotDefined');
+    }
+
+    public function testItReturnsExpectedTypeWithEnums(): void
+    {
+        $map = new GuidTypeMap();
+        $map->define(TestBackedEnum::Foo, 'SomeFooType');
+        $map->define(TestBackedEnum::Bar, TestBackedIntEnum::FooBar);
+        $map->define(TestUnitEnum::Baz, 'SomeBazType');
+        $map->define(TestUnitEnum::Bat, TestBackedIntEnum::BazBat);
+
+        $id = Uuid::random();
+
+        $this->assertSame('SomeFooType', $map->typeFor(TestBackedEnum::Foo));
+        $this->assertObjectEquals(Guid::fromUuid('SomeFooType', $id), $map->guidFor(TestBackedEnum::Foo, $id));
+        $this->assertSame(TestBackedIntEnum::FooBar, $map->typeFor(TestBackedEnum::Bar));
+        $this->assertObjectEquals(Guid::fromUuid(TestBackedIntEnum::FooBar, $id), $map->guidFor(TestBackedEnum::Bar, $id));
+        $this->assertSame('SomeBazType', $map->typeFor(TestUnitEnum::Baz));
+        $this->assertObjectEquals(Guid::fromUuid('SomeBazType', $id), $map->guidFor(TestUnitEnum::Baz, $id));
+        $this->assertSame(TestBackedIntEnum::BazBat, $map->typeFor(TestUnitEnum::Bat));
+        $this->assertObjectEquals(Guid::fromUuid(TestBackedIntEnum::BazBat, $id), $map->guidFor(TestUnitEnum::Bat, $id));
     }
 }